move strings to pkg/util/strings
