+++ pkg/util/strings/strings_test.go
@@ -110,125 +110,3 @@ func TestNormalizeURL(t *testing.T) {
 		require.EqualValues(t, test.expected, normalized)
 	}
 }
-
-func TestTable(t *testing.T) {
-	tests := []struct {
-		header   Row
-		body     []Row
-		expected []FormattedRow
-	}{
-		{
-			header: Row{
-				Content: []string{
-					"FOO-HEADER",
-					"BAR-HEADER",
-				},
-			},
-			body: []Row{
-				Row{
-					Content: []string{
-						"foo",
-						"bar",
-					},
-				},
-				Row{
-					Content: []string{
-						"another-foo",
-						"another-bar",
-					},
-				},
-			},
-			expected: []FormattedRow{
-				FormattedRow{
-					Content: "FOO-HEADER  BAR-HEADER",
-				},
-				FormattedRow{
-					Content: "==========  ==========",
-				},
-				FormattedRow{
-					Content: "foo         bar",
-				},
-				FormattedRow{
-					Content: "another-foo another-bar",
-				},
-			},
-		},
-	}
-	for _, test := range tests {
-		formatted, err := Table(test.header, test.body)
-		if err != nil {
-			t.Error(err)
-		}
-		require.EqualValues(t, test.expected, formatted)
-	}
-}
-
-func TestPadRows(t *testing.T) {
-	tests := []struct {
-		input    [][]string
-		expected string
-	}{
-		{
-			input:    [][]string{},
-			expected: ``,
-		},
-		{
-			input: [][]string{
-				[]string{"Hello", "World"},
-			},
-			expected: "Hello World\n",
-		},
-		{
-			input: [][]string{
-				[]string{"Hello", "World"},
-				[]string{"Hi", "World"},
-			},
-			expected: `Hello World
-Hi    World
-`,
-		},
-		{
-			input: [][]string{
-				[]string{"Hello"},
-				[]string{"Hi", "World"},
-			},
-			expected: `Hello
-Hi    World
-`,
-		},
-		{
-			input: [][]string{
-				[]string{},
-				[]string{"Hi", "World"},
-			},
-			expected: `
-Hi World
-`,
-		},
-		{
-			input: [][]string{
-				[]string{"Hello", "World"},
-				[]string{""},
-			},
-			expected: `Hello World
-
-`,
-		},
-		{
-			input: [][]string{
-				[]string{""},
-				[]string{""},
-			},
-			expected: `
-
-`,
-		},
-	}
-	for _, test := range tests {
-		padded, err := PadRows(test.input)
-		if err != nil {
-			t.Error(err)
-		}
-		require.EqualValues(t, test.expected, padded)
-	}
-}
